feat: Update FreeRTOS kernel and integrate new firmware components for RPC, queues, events, and task monitoring.
