+++ firmware/config/firmware_config.h
@@ -1,0 +1,13 @@
+#ifndef FIRMWARE_CONFIG_H
+#define FIRMWARE_CONFIG_H
+
+#define CLICK_QUEUE_LEN 32
+#define DISPLAY_PERIOD_MS 150
+#define MONITOR_PERIOD_MS 5000
+#define RPC_POLL_PERIOD_MS 100
+#define SCORES_REFRESH_MS 5000
+#define RPC_REGISTER_RETRY_MS 2000
+#define TURBO_DURATION_MS 10000
+#define MILESTONE_GLOW_TICKS 10
+
+#endif /* FIRMWARE_CONFIG_H */
