+++ firmware/config/net_config.h
@@ -1,0 +1,12 @@
+#ifndef NET_CONFIG_H
+#define NET_CONFIG_H
+
+#ifndef WIFI_SSID
+#define WIFI_SSID ""
+#endif
+
+#ifndef WIFI_PASSWORD
+#define WIFI_PASSWORD ""
+#endif
+
+#endif /* NET_CONFIG_H */
