+++ firmware/middleware/game_events.h
@@ -1,0 +1,6 @@
+#ifndef GAME_EVENTS_H
+#define GAME_EVENTS_H
+
+void trigger_milestone_feedback(void);
+
+#endif /* GAME_EVENTS_H */
