+++ firmware/tasks/task_buttons.h
@@ -1,0 +1,6 @@
+#ifndef TASK_BUTTONS_H
+#define TASK_BUTTONS_H
+
+void task_buttons(void *param);
+
+#endif /* TASK_BUTTONS_H */
