+++ firmware/tasks/task_display.h
@@ -1,0 +1,6 @@
+#ifndef TASK_DISPLAY_H
+#define TASK_DISPLAY_H
+
+void task_display(void *param);
+
+#endif /* TASK_DISPLAY_H */
