+++ firmware/tasks/task_monitor.h
@@ -1,0 +1,6 @@
+#ifndef TASK_MONITOR_H
+#define TASK_MONITOR_H
+
+void task_monitor(void *param);
+
+#endif /* TASK_MONITOR_H */
