+++ firmware/tasks/task_rpc.h
@@ -1,0 +1,6 @@
+#ifndef TASK_RPC_H
+#define TASK_RPC_H
+
+void task_rpc(void *param);
+
+#endif /* TASK_RPC_H */
